+++ src/main/resources/application.properties
@@ -1,4 +1,3 @@
 db.url=jdbc:h2:./janggi;AUTO_SERVER=TRUE;INIT=RUNSCRIPT FROM 'classpath:/init.sql'
 db.username=sa
 db.password=
-# ?? ???? ?????, .gitignore? ??
