+++ src/main/resources/init.sql
@@ -1,16 +1,15 @@
-DROP TABLE IF EXISTS Piece_Position;
-DROP TABLE IF EXISTS Game_Room;
+DROP TABLE IF EXISTS piece_position;
+DROP TABLE IF EXISTS game_room;
 
-CREATE TABLE Game_Room
+CREATE TABLE game_room
 (
     janggi_game_id INT AUTO_INCREMENT PRIMARY KEY,
     room_name      VARCHAR(10) NOT NULL,
-    game_status    VARCHAR(10) NOT NULL CHECK (game_status IN ('PLAYING', 'FINISHED')),
     last_turn      VARCHAR(5)  NOT NULL CHECK (last_turn IN ('CHO', 'HAN')),
     last_played_at DATE        NOT NULL
 );
 
-CREATE TABLE Piece_Position
+CREATE TABLE piece_position
 (
     piece_position INT AUTO_INCREMENT PRIMARY KEY,
     janggi_game_id INT         NOT NULL,
@@ -23,7 +22,7 @@ CREATE TABLE Piece_Position
 
     CONSTRAINT fk_game
         FOREIGN KEY (janggi_game_id)
-            REFERENCES Game_Room (janggi_game_id),
+            REFERENCES game_room (janggi_game_id),
 
     CONSTRAINT uq_position
         UNIQUE (janggi_game_id, piece_row, piece_column)
